Hunger and food

+++ core/src/org/rexellentgames/dungeon/util/MathUtils.java
@@ -1,7 +1,7 @@
 package org.rexellentgames.dungeon.util;
 
 public class MathUtils {
-	public static int clamp(int min, int max, int val) {
+	public static float clamp(float min, float max, float val) {
 		return Math.max(min, Math.min(max, val));
 	}
 }
